修改了获取积分记录时候的null bug

+++ src/main/java/com/lst/lc/dao/LetterDao.java
@@ -18,6 +18,6 @@ public interface LetterDao {
 	 */
 	public List<Letter> getAll(int uid);
 	
-	public long getUnRead(int uid);
+	public int getUnRead(int uid);
 	
 }
